+++ src/main/java/me/pajic/mapstitch/worldmap/WorldMapScreen.java
@@ -11,6 +11,7 @@ import me.pajic.mapstitch.component.ModDataComponents;
 import me.pajic.mapstitch.config.ModConfigHolder;
 import me.pajic.mapstitch.item.ModItems;
 import me.pajic.mapstitch.keybind.ModKeybinds;
+import me.pajic.mapstitch.networking.payload.C2SEjectMap;
 import me.pajic.mapstitch.util.CompatFlags;
 import me.pajic.mapstitch.util.ModUtil;
 import net.minecraft.client.Minecraft;
@@ -21,6 +22,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.sounds.SoundEvents;
 import net.minecraft.util.FastColor;
 import net.minecraft.util.Mth;
 import net.minecraft.world.item.ItemStack;
@@ -28,6 +30,7 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.item.component.BundleContents;
 import net.minecraft.world.level.saveddata.maps.MapId;
 import net.minecraft.world.level.saveddata.maps.MapItemSavedData;
+import net.neoforged.neoforge.network.PacketDistributor;
 import org.apache.commons.lang3.text.WordUtils;
 import org.jetbrains.annotations.NotNull;
 import org.joml.Vector2i;
@@ -54,6 +57,7 @@ public class WorldMapScreen extends Screen {
 	private int mapSize;
 	private double mapPixels;
 	private double camX, camZ;
+	private int mouseX, mouseY;
 	private int zoomLevel;
 	private float zoom;
 	private int screenW, screenH;
@@ -79,10 +83,18 @@ public class WorldMapScreen extends Screen {
 	private record MapDataWithId(MapId id, MapItemSavedData data) {}
 
 	private record GridPos(int gx, int gy, int s) {}
+
 	private GridPos worldToGrid(int worldX, int worldZ, int scale) {
 		return new GridPos(
 				Math.floorDiv(worldX, 128 << scale),
 				Math.floorDiv(worldZ, 128 << scale),
+				scale
+		);
+	}
+	private GridPos screenToGrid(double screenX, double screenY, int scale) {
+		return new GridPos(
+				Math.floorDiv(Mth.floor(screenToWorldX(screenX)) + 64, 128 << scale),
+				Math.floorDiv(Mth.floor(screenToWorldZ(screenY)) + 64, 128 << scale),
 				scale
 		);
 	}
@@ -93,13 +105,15 @@ public class WorldMapScreen extends Screen {
 	private double worldToScreenX(double worldX) { return (worldX - camX) * zoom + screenW / 2.0; }
 	private double worldToScreenZ(double worldZ) { return (worldZ - camZ) * zoom + screenH / 2.0; }
 	private double screenToWorldX(double screenX) { return (screenX - screenW / 2.0) / zoom + camX; }
-	private double screenToWorldZ(double screenZ) { return (screenZ - screenH / 2.0) / zoom + camZ; }
+	private double screenToWorldZ(double screenY) { return (screenY - screenH / 2.0) / zoom + camZ; }
 
 	@Override
 	public void render(@NotNull GuiGraphics graphics, int mouseX, int mouseY, float partialTick) {
 		super.render(graphics, mouseX, mouseY, partialTick);
 		mapSize = 128 << scale;
 		mapPixels = mapSize * zoom;
+		this.mouseX = mouseX;
+		this.mouseY = mouseY;
 		int highlightColor = ModConfigHolder.options().worldMapTextHighlightColor.color;
 		int xBoundMin = Math.floorDiv((int)(camX - (screenW/2d)/zoom), mapSize);
 		int xBoundMax = Math.floorDiv((int)(camX + (screenW/2d)/zoom), mapSize) + 1;
@@ -211,6 +225,7 @@ public class WorldMapScreen extends Screen {
 				new ObjectBooleanImmutablePair<>(Component.translatable("mapstitch.gui.worldmap.help", Component.translatable("mapstitch.gui.worldmap.help_key").withColor(c)), ModConfigHolder.options().worldMapHelp && !help),
 				new ObjectBooleanImmutablePair<>(Component.translatable("mapstitch.gui.worldmap.help_control", Component.translatable("mapstitch.gui.worldmap.help_key").withColor(c)), help),
 				new ObjectBooleanImmutablePair<>(Component.translatable("mapstitch.gui.worldmap.exit_control", Component.translatable("mapstitch.gui.worldmap.exit_key").withColor(c), Component.keybind(ModKeybinds.OPEN_WORLD_MAP.getName()).withColor(c)), help),
+				new ObjectBooleanImmutablePair<>(Component.translatable("mapstitch.gui.worldmap.eject_control", Component.translatable("mapstitch.gui.worldmap.eject_key").withColor(c)), help),
 				new ObjectBooleanImmutablePair<>(Component.translatable("mapstitch.gui.worldmap.grid_control", Component.translatable("mapstitch.gui.worldmap.grid_key").withColor(c)), help),
 				new ObjectBooleanImmutablePair<>(Component.translatable("mapstitch.gui.worldmap.follow_control", Component.translatable("mapstitch.gui.worldmap.follow_key").withColor(c)), help),
 				new ObjectBooleanImmutablePair<>(Component.translatable("mapstitch.gui.worldmap.scale_control", Component.translatable("mapstitch.gui.worldmap.scale_key").withColor(c)), help),
@@ -316,6 +331,15 @@ public class WorldMapScreen extends Screen {
 		}
 	}
 
+	@SuppressWarnings("DataFlowIssue")
+	private void ejectMapAtCursor() {
+		MapDataWithId mapDataWithId = MAPS.get(screenToGrid(mouseX, mouseY, scale));
+		if (mapDataWithId != null) {
+			PacketDistributor.sendToServer(new C2SEjectMap(mapDataWithId.id));
+			MC.player.playSound(SoundEvents.BUNDLE_REMOVE_ONE);
+		}
+	}
+
 	@Override
 	public boolean mouseDragged(double mouseX, double mouseY, int button, double dragX, double dragY) {
 		if (button == 0) {
@@ -372,6 +396,10 @@ public class WorldMapScreen extends Screen {
 		}
 		if (ModKeybinds.OPEN_WORLD_MAP.matches(keyCode, scanCode)) {
 			onClose();
+			return true;
+		}
+		if (modifiers == 2 && keyCode == InputConstants.KEY_Q) {
+			ejectMapAtCursor();
 			return true;
 		}
 		return super.keyPressed(keyCode, scanCode, modifiers);